[iOS] refactor : HTTPURLResponse을 사용하여 validate
Associate with #34
