+++ mobile/TodoList/TodoList/Networking/NetworkManager.swift
@@ -8,6 +8,16 @@
 
 import Foundation
 
+enum HTTPError: Error {
+   case badRequest
+   
+   var localizedDescription: String {
+       switch self {
+       case .badRequest: return "Bad Request"
+       }
+   }
+}
+
 class NetworkManager {
     private let session: URLSession
     
@@ -33,21 +43,31 @@ class NetworkManager {
         }.resume()
     }
     
-    func requestNewCard(card: Card, completion: @escaping (Result<Response, Error>) -> Void) {
+    func requestNewCard(card: Card, completion: @escaping (Result<CardIDResponse, Error>) -> Void) {
         guard let request = APIRouter.newCard(card: card).urlRequest else { return }
         
-        session.dataTask(with: request) { data, _, error in
+        session.dataTask(with: request) { data, response, error in
             if let error = error {
                 completion(.failure(error))
                 return
             }
-            guard let data = data else { return }
+            guard let data = data, let httpResponse = response as? HTTPURLResponse else { return }
+            if !httpResponse.isValid() { completion(.failure(HTTPError.badRequest)) }
             do {
                 let response = try JSONDecoder().decode(CardIDResponse.self, from: data)
-                response.validate(statusCode: .ok) { completion($0) }
+                completion(.success(response))
             } catch {
                 completion(.failure(error))
             }
         }.resume()
+    }
+}
+
+private extension HTTPURLResponse {
+    func isValid() -> Bool {
+        switch statusCode {
+        case 200..<300: return true
+        default: return false
+        }
     }
 }
